+++ docs/ReleaseNotes.md
@@ -1,5 +1,13 @@
 # ReleaseNotes
 
+## [2.0.0](https://www.nuget.org/packages/DbTool.Core/2.0.1)
+
+- Add support for new C# features
+  - Global using
+  - File-scoped namespace
+  - Nullable reference types
+- Support for custom indentation, global usings
+
 ## [1.1.0](https://www.nuget.org/packages/DbTool.Core/1.1.0)
 
 - Initial plugin based release